CMPE-244: HW2: Test external LED and Switch via GPIOs
This patch simply adds a new command ('led') to do a GPIO test for
CMPE-244 homework 2.

The routine of this command listens a GPIO (default: P0.30) from a
external switch so as to complete different LED lighting effects
through another GPIO (default: P2.7) -- If switch-on, blink the LED;
Otherwise, just turn it on.

It also allows users to configure the total time and blinking period
along with desired GPIO ports.

Signed-off-by: Nicolin Chen <nicoleotsuka@gmail.com>