CMPE-244: HW7: Test semaphore with interrupt and timer
This patch adds a new command to test the semaphore with interrupt
and timer supports. The interrupt receives a falling edge trigger
from external GPIO0.30 (default) and starts a timer for its switch
debounce. The timer delays 300ms for the debounce task and then
give a semaphore to a pending task.

This patch also enables the timer function in the FreeRTOS since
this assignment needs us to use it for switch debounce.

Signed-off-by: Nicolin Chen <nicoleotsuka@gmail.com>